+++ ouvidoria/domains/compliments.py
@@ -6,7 +6,17 @@ formatter = Formatter()
 
 
 class Compliments:
+    ideasIds = []
+    ideasTypes = []
+    ideasAuthors = []
+    ideasDescriptions = []
+
     def getCompliments(self):
+        self.complimentsIds = []
+        self.complimentsTypes = []
+        self.complimentsAuthors = []
+        self.complimentsDescriptions = []
+
         connection = openDatabase("localhost", "root", "root", "db_ouvidoria")
 
         formatter.header("Elogios")
@@ -19,9 +29,17 @@ class Compliments:
             )
         )
         for itemDB in listCompliments:
+            self.complimentsIds.append(itemDB[0])
+            self.complimentsTypes.append(itemDB[1])
+            self.complimentsAuthors.append(itemDB[2])
+            self.complimentsDescriptions.append(itemDB[3])
+        for i in range(len(self.complimentsIds)):
             print(
                 "{:<12} {:<15} {:<15} {:<10}".format(
-                    str(itemDB[0]), itemDB[1], itemDB[2], itemDB[3]
+                    i + 1,
+                    self.complimentsTypes[i],
+                    self.complimentsAuthors[i],
+                    self.complimentsDescriptions[i],
                 )
             )
 
@@ -34,7 +52,7 @@ class Compliments:
         data = ("Elogio", author, compliment)
         insertInDatabase(connection, sqlCode, data)
 
-        print("\nElogio adicionado com sucesso!")
+        formatter.successEmitter("Elogio adicionado com sucesso!")
 
         closeDatabase(connection)
 
@@ -45,21 +63,30 @@ class Compliments:
 
         idOfCompliment = validate.readInt("\nNúmero do elogio: ")
 
-        sqlCode = "DELETE FROM feedbacks WHERE id = %s"
-        data = (idOfCompliment,)
-        deleteRegistryInDatabase(connection, sqlCode, data)
+        if idOfCompliment == 0 or idOfCompliment > len(self.complimentsIds):
+            formatter.errorEmitter("Elogio não encontrado!")
+        else:
+            id = self.complimentsIds[idOfCompliment - 1]
+            sqlCode = "DELETE FROM feedbacks WHERE id = %s"
+            data = (id,)
+            deleteRegistryInDatabase(connection, sqlCode, data)
 
-        print("\nElogio removido com sucesso!\n")
+            formatter.successEmitter("Elogio removido com sucesso!")
 
         closeDatabase(connection)
 
-    def editCompliment(self, id: int, newCompliment: str):
+    def editCompliment(self, idOfCompliment: int):
         connection = openDatabase("localhost", "root", "root", "db_ouvidoria")
 
-        sqlCode = "UPDATE feedbacks SET feedback = %s WHERE id = %s"
-        data = (newCompliment, id)
-        updateDatabase(connection, sqlCode, data)
+        if idOfCompliment == 0 or idOfCompliment > len(self.complimentsIds):
+            formatter.errorEmitter("Elogio não encontrado!")
+        else:
+            newCompliment = validate.readString("\nDigite seu novo elogio aqui:\n")
+            id = self.complimentsIds[idOfCompliment - 1]
+            sqlCode = "UPDATE feedbacks SET feedback = %s WHERE id = %s"
+            data = (newCompliment, id)
+            updateDatabase(connection, sqlCode, data)
 
-        print("\nElogio editado com sucesso!\n")
+            formatter.successEmitter("Elogio editado com sucesso!")
 
         closeDatabase(connection)
